% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{length2weight}
\alias{length2weight}
\title{Calculate the weight of a fish from its length using a \eqn{W = a ∗ L^b} model}
\usage{
length2weight(length, alpha, beta)
}
\arguments{
\item{length}{numeric The length of the fish in cm}

\item{alpha}{numeric The alpha coefficient}

\item{beta}{numeric The beta coefficient}
}
\value{
The weight of the fish in kg
}
\description{
This function calculates the weight of a fish from its length using
\eqn{W = a ∗ L^b} model. The coefficients alpha and beta are available for several
species in the data frame \link{species_data} and can be retrieved using the
\code{\link[=species_coeff]{species_coeff()}} function.
}
\examples{
perch <- species_coeff("ABBO")
length2weight(50, perch$alpha, perch$beta)
pike <- species_coeff(name ="Gädda")
length2weight(50, pike$alpha, pike$beta)

}
