% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{slurp_excel}
\alias{slurp_excel}
\title{Read many excel files and combine them into a single data frame}
\usage{
slurp_excel(fnames, ...)
}
\arguments{
\item{fnames}{list of file names}

\item{...}{optional parameters passed to readxl::read_excel}
}
\value{
A data frame with the contents of all the files
}
\description{
Combine many excel files into a single data frame. All files must have the
same structure. No error checking is done and weird things will happen you
break this assumption.
}
\examples{
\dontrun{
fnames <- c("file1.xlsx", "file2.xlsx", "file3.xlsx")
df <- slurp_excel(fnames)
}
}
