% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{species_coeff}
\alias{species_coeff}
\title{Get coefficients for a species to calculate the weight from the length}
\usage{
species_coeff(code = NULL, name = NULL)
}
\arguments{
\item{code}{character The species code (e.g. "ABBO")}

\item{name}{character The species name (e.g. "Abborre")}
}
\value{
A data frame with the alpha and beta coefficients for the species.
}
\description{
This function returns pre-calulated the coefficients for a species uses to
estimate weight from the length. The coefficients are available in the
data frame \code{demop::species_data}. You can provide either the species code
or the species name (but not both) to find the data.
}
\examples{
species_coeff("ABBO")

}
