% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_documentation.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{A data frame with 8 columns containing various species related constants}
\format{
A data frame with 147 rows and 8 columns
\describe{
\item{code}{Short code for species}
\item{name}{Species name}
\item{alpha}{Used to estimate weight from length}
\item{beta}{Used to estimate weight from length}
\item{min_length}{Minimum length used in estimation of alpha and beta?}
\item{max_length}{Maximum length used in estimation of alpha and beta?}
\item{min_weight}{Minimum weight used in estimation of alpha and beta?}
\item{max_weight}{Maximum weight used in estimation of alpha and beta?}
}
}
\usage{
species_data
}
\description{
Data set containing various species related constants used in Spöreg
}
\keyword{datasets}
